Moved getrolesofuseringuild function out of test menu and made a function for changing a menu to staff menu
